+++ ashley/src/main/kotlin/ktx/ashley/mappers.kt
@@ -8,8 +8,9 @@ import com.badlogic.ashley.core.ComponentMapper
  *
  * Provides `O(1)` retrieval of [Component]s for an [com.badlogic.ashley.core.Entity].
  *
- * @param T the [Component] type to create a [ComponentMapper] for
- * @return the [ComponentMapper]
+ * @param T the [Component] type to create a [ComponentMapper] for.
+ * @return a  [ComponentMapper]
  * @see ComponentMapper
+ * @see Component
  */
-inline fun <reified T: Component> mapperFor(): ComponentMapper<T> = ComponentMapper.getFor(T::class.java)
+inline fun <reified T : Component> mapperFor(): ComponentMapper<T> = ComponentMapper.getFor(T::class.java)